add some functions

+++ src/include/breakpoint.h
@@ -7,9 +7,9 @@
 
 
 typedef struct breakpoint {
-	long 	orig_data;
-	long 	address;
-	int 	enabled;
+	unsigned long 	orig_data;
+	long 		address;
+	int 		enabled;
 } breakpoint_t;
 
 
@@ -18,8 +18,8 @@ int 		breakpoint_compare(void* address, void* list_node);
 void 		breakpoints_init();
 breakpoint_t*	breakpoint_create(pid_t dbpid, void* value);
 
-int 		breakpoint_enable(pid_t dbpid, long address);
-int 		breakpoint_disable(pid_t dbpid, long address);
+int 		breakpoint_enable(pid_t dbpid, long long address);
+int 		breakpoint_disable(pid_t dbpid, long long address);
 void 		step_to_breakpoint(pid_t pid);
